// src/components/condition-order-table/types.ts
export type ConditionOrderKind = 'condition' | 'tp_sl'
export type ConditionOrderLifecycle = 'current' | 'history'
// ...
export interface ConditionOrderListQuery {
  end: string
  keyword: string
  kind: ConditionOrderKind
  lifecycle: ConditionOrderLifecycle
  order_type: boolean | number | string
  page_no: number
  page_size: number
  reduce_only: boolean | number | string
  side: boolean | number | string
  start: string
  sym: string
  user_type?: boolean | number | string
}